refactor:用户管理模块
重构路由的引用，和代码中的问题
